Separate avatar owner abstraction logic
Getting an avatar's owner from the `$id_or_email` parameter can be a
repetitive task. Simplifying this resulted in a separated piece of logic
that determines the avatar's owner, while on the other side making sure
every abstraction is saved in cache and made available when needed. This
makes for a more streamlined getting-the-owner logic flow.