Fix a bug where all avatars might be identified as the sample avatar
To treat the sample avatar in the admin settings differently, the 'avatar_defaults' filter was used to identify the sample avatar. A side effect of this approach was that all avatars would be treated as the sample avatar on all pages or WordPress requests with this filter. Naturally this shouldn't occur, but it does for sites using BuddyPress 8.0+ where the 'avatar_defaults' filter is run for every avatar in `bp_core_is_default_avatar`. To mitigate this the identifying of sample avatars is moved to a new location with a condition that checks for the Discussion settings page.